the unfilled lines are now replaced with the original pixels
Im not sure if this is better. I now take the original image, copy that and then write over it, and this makes the lines in the corners now present, but not pixelated. If its a small line it looks fine. But if its pixel_size - 1 then its pretty clearly there. I am debating on if just slightly downslcaling the image would be best...